refactor(mobile): make the switcher's shared row actions artifact-only
Terminal rows drive their own lifetime-gated menu, so nothing reaches the
shared component as a terminal any more: its kind union, the "Close" label,
the immediate-close branch and the null confirm description were all
unreachable, and the hook behind it mounted a terminal-kill mutation on
every agent and artifact row that would never fire.

Narrow both to what is left. The component becomes SwitcherArtifactRowActions,
matching the SwitcherAgentRowActions / SwitcherTerminalRowActions pair it now
sits beside, and drops the leading-entries indirection that existed only to
keep the export mutation away from non-artifact rows - there are none.
